getting competency names now has its own api and service
