+++ server/src/service/application_service.ts
@@ -1,5 +1,4 @@
 import {Database} from '../integration/database';
-import {CompetenceDAO} from '../integration/competence_dao';
 import {fullApplicationDTO} from '../model/dto/full_application_dto';
 import {AvailabilityDAO} from '../integration/availability_dao';
 import {CompetenceProfileDAO} from '../integration/competence_profile_dao';
@@ -8,19 +7,6 @@ import {ConflictError} from '../utilities/custom_errors';
 
 export class ApplicationService {
   constructor() {}
-
-  async getCompetencies() {
-    const db = Database.getInstance().database;
-    try {
-      return await db.transaction(async transaction => {
-        const competenceDAO = CompetenceDAO.getInstance();
-        return await competenceDAO.getAllCompetencies(transaction);
-      });
-    } catch (error) {
-      throw error;
-    }
-  }
-
   async handleApplication(application: fullApplicationDTO) {
     const db = Database.getInstance().database;
     try {
